Updated Random number generator to take a floor value.
- Changed IEntities UUIDs to be uint64_t.

+++ Humble/src/Core/IEntity.h
@@ -7,7 +7,7 @@ namespace HBL {
 
 	struct HBL_API IEntity
 	{
-		uint32_t m_UUID = 0x0000;
+		uint64_t m_UUID = 0x0000;
 	};
 
 }
